Fix some compiler warnings
This commit changes code so as not to emit compiler warnings, when
most warnings are enabled. Most of this involves changing old-style
casts like (double*)x to new-style static_cast<>s, and removing
unnecessary consts.

+++ src/dss_pinlo/dss_pinlo_test.cpp
@@ -9,8 +9,8 @@ int main(int argc, char** argv) {
     DSSpiNLO ffs("PINLO.DAT");
     double zlist[] = {exp(-4.605), exp(-3.912), 0.8};
     double qs2list[] = {exp(0), exp(0.2231), 3.91213};
-    for (int iz = 0; iz < sizeof(zlist) / sizeof(zlist[0]); iz++) {
-        for (int iq = 0; iq < sizeof(qs2list) / sizeof(qs2list[0]); iq++) {
+    for (size_t iz = 0; iz < sizeof(zlist) / sizeof(zlist[0]); iz++) {
+        for (size_t iq = 0; iq < sizeof(qs2list) / sizeof(qs2list[0]); iq++) {
             double z = zlist[iz];
             double qs2 = qs2list[iq];
             ffs.update(z, qs2);
